feat(prop): allow deep paths (#1158)
Closes: #830

Extending `prop` to accept a variadic path of props into an
object/array.

The prop names are inferred one after the other, allowing typescript IDE
typeahead to suggest the possible props based on the previous props
already in the path.

Types support arrays, tuples, and unions of objects correctly, and
allows selecting props even when they don't exist on all parts of the
union.

I'm planning on sending a very simple `defaultTo` utility that would
provide a stricter version of the coelsece operator, this would allow us
to deprecate `pathOr` which has gotten multiple issues opened for it
since it was added and switch users to a composite of `prop` and
`defaultTo`.