feat(defaultTo): introduce `defaultTo` (#1159)
As a follow up to #1158 we needed an easy way to utilize `prop` to
replace `pathOr`. This meant adding an additional utility that could be
easily changed after it to provide a fallback.

In order for this utility to have merit on it's own, and provide some
typing magic beyond just wrapping a built-in operator, I built it in a
way which enforces it to be as strict as possible.

It does this by limiting the scope for both the data param, and the
fallback param, so that data must be nullish, and fallback must match
it's type.

---------

Co-authored-by: CJ Quines <cjquines0@gmail.com>