feat: migrate build from tsup to tsdown (#1172)
* Migrate from tsup to tsdown. tsdown is a drop-in replacement. We
needed to do some minor changes, though:
- Supports globs for the entryPoint, no need for manually building the
list.
  - doesn't support the `splitting` setting for cjs.
  - `clean` is enabled by default.
* Add source maps (as requested in issue #1050)
* Integrate attw and publint directly in the build, drop the dedicated
lint stage from CI
* Update the CI to better handle building (because tsdown doesn't offer
a `--dts-only` flag to keep it efficient) by doing the build once,
caching it, and then restoring it separately for each ts version in the
matrix. We do this for the canary and ci pipelines.
* Install `lightningcss` into the docs project, instead of relying on it
being installed via dependencies. npm decided to change it to an
optional peer dependency, and without installing it explicitly, astro
fails to build when it's missing.
* Test types on TypeScript 5.9!
* Pinned action versions to patch version for consistent repros

Closes: #1050