docs: fix click surface for migration dropdown links (#1178)
The click area of the links in the migrations dropdown doesn't match
it's visual space as well as its hover-state. Making it easy to miss the
link, but the dropdown still closes, so it feels weird.

Example:
<img width="242" height="233" alt="image"
src="https://github.com/user-attachments/assets/e541b333-8dab-431b-b128-559027eeb302"
/>

This makes the dropdown element be the anchor itself, so it takes all
the space, and matches the visual size.