chore: bump all deps manually (october 2025) (#1211)
The dependabot PRs can't merge cleanly because of some breaking changes,
even at the patch level. As is common in these monthly bumps, it ends up
being easier to do a manual PR for all bumps while fixing any issues.

Many of the changes in this PR result from trying to update to type-fest
v5. However, I eventually opted not to update it because it defines a
minimum requirement of node 20, which is above our defined minimum for
v2 (of 18). We will have to drop CommonJS support in a future major
version to support type-fest v5+

Additionally:
- Allow the pkg-pr-new task to run even when tests/lint fail, so the
build could be used manually to debug failing tests, and without
handling all the quality concerns upfront.
- modernize algolia docsearch api
- Remove `lightningcss` and release the minor version lock from our node
version (via `.nvmrc`) because the underlying issue that was encountered
previously has now been resolved.
- Fix toCamelCase default option typing (it was broken) by extracting
the logic from toTitleCase and reusing it in both places
- lint fixes
- Minor changes and fixes to `groupByProp` to support a future bump of
type-fest to v5.
- add funding to package.json because we can use funding.