feat(build): generate a monolithic type definition file (#1241)
Fixes: #1175

Our current build genreates tiny files that are good for tree-shaking
but not good for reexporting our types when used within a multi-output
project.

To fix this I needed to restructure our build process to allow these 2
different aspects to play along nicely.

NOTE: I'm using a `feat` PR type instead of a `fix` although the change
is largely under-the-hood and shouldn't cause any breakages. This is
just in case, so that user get a proper "heads up" that our build
artifacts changed.